substantial rework: enable caching of downloaded schedules
This commit makes it so that the engine caches downloaded
schedules in .kde/share/apps/plasma_engine_rtddenver. The
engine is smart enough to load a schedule once a day to make
sure that the cached schedules are still valid.

The source syntax has changed. It is now
  Routes
  ValidAsOf
  DirectionOf <route>
  ScheduleOf <route> <direction>

TODO: add sourceAdded() emission at the appropriate place(s)